+++ src/lib/api/authService.ts
@@ -1,3 +1,4 @@
+import { Role, User } from "@/types/types";
 import instance from "./instance";
 
 export function postLogin(data: { email: string; password: string }): Promise<{
@@ -12,7 +13,7 @@ export function postLogin(data: { email: string; password: string }): Promise<{
 	}
 }
 
-export function postSignUp(data: { name: string; email: string; phone: string; password: string }): Promise<{
+export function postSignUp(data: { name: string; email: string; phone: string; password: string; role: Role }): Promise<{
 	user: User;
 	accessToken: string;
 	refreshToken: string;
