+++ src/pages/login.tsx
@@ -43,7 +43,7 @@ function LogIn() {
   };
 
   return (
-    <main className="flex flex-col justify-center items-center gap-[32px] h-screen w-[384px] max-w-full mx-auto p-[4px]">
+    <main className="flex flex-col justify-center items-center gap-[32px] w-[384px] max-w-full mx-auto p-[4px] my-[64px]">
       <Image src={logo_xl} alt="Logo" width={384} height={124} priority />
       <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-[16px] items-stretch w-full">
         <label className="w-full text-lg" htmlFor="email">이메일</label>
